# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: paperz.pk — school exam paper maker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [session]
      summary: Get current authenticated user
      description: Returns the current user with role and school, provisioning a local record if needed.
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /onboarding/school:
    post:
      operationId: onboardSchool
      tags: [session]
      summary: Create a school during onboarding
      description: Creates a new school and assigns the current user as its school admin, seeding starter content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List and filter questions
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: chapterId
          in: query
          schema:
            type: integer
        - name: topicId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/QuestionType"
        - name: medium
          in: query
          schema:
            $ref: "#/components/schemas/Medium"
        - name: difficulty
          in: query
          schema:
            $ref: "#/components/schemas/Difficulty"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
  /questions/import:
    post:
      operationId: importQuestions
      tags: [questions]
      summary: Bulk import questions from parsed CSV/Excel rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Role:
      type: string
      enum: [super_admin, school_admin, teacher]
    QuestionType:
      type: string
      enum: [mcq, short, long, exercise, conceptual, past_paper]
    Medium:
      type: string
      enum: [urdu, english, dual]
    Difficulty:
      type: string
      enum: [easy, medium, hard]
    Me:
      type: object
      required: [authenticated, needsOnboarding]
      properties:
        authenticated:
          type: boolean
        needsOnboarding:
          type: boolean
        id:
          type: ["integer", "null"]
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          oneOf:
            - $ref: "#/components/schemas/Role"
            - type: "null"
        schoolId:
          type: ["integer", "null"]
        school:
          oneOf:
            - $ref: "#/components/schemas/School"
            - type: "null"
    School:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    SchoolInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        phone:
          type: string
        logoUrl:
          type: string
...
    Question:
      type: object
      required: [id, classId, subjectId, type, medium, difficulty, marks, text]
      properties:
        id:
          type: integer
        classId:
          type: integer
        subjectId:
          type: integer
        chapterId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        type:
          $ref: "#/components/schemas/QuestionType"
        medium:
          $ref: "#/components/schemas/Medium"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        marks:
          type: integer
        text:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        answer:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        subjectName:
          type: ["string", "null"]
        chapterName:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    QuestionInput:
      type: object
      required: [classId, subjectId, type, medium, difficulty, marks, text]
      properties:
        classId:
          type: integer
        subjectId:
          type: integer
        chapterId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        type:
          $ref: "#/components/schemas/QuestionType"
        medium:
          $ref: "#/components/schemas/Medium"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        marks:
          type: integer
          minimum: 0
        text:
          type: string
          minLength: 1
        options:
          type: array
          items:
            type: string
        answer:
          type: string
    QuestionUpdate:
      type: object
      properties:
        chapterId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        type:
          $ref: "#/components/schemas/QuestionType"
        medium:
          $ref: "#/components/schemas/Medium"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        marks:
          type: integer
          minimum: 0
        text:
          type: string
          minLength: 1
        options:
          type: array
          items:
            type: string
        answer:
          type: string
    QuestionImportInput:
      type: object
      required: [classId, subjectId, rows]
      properties:
        classId:
          type: integer
        subjectId:
          type: integer
        chapterId:
          type: ["integer", "null"]
        rows:
          type: array
          items:
            $ref: "#/components/schemas/QuestionImportRow"
    QuestionImportRow:
      type: object
      required: [text]
      properties:
        text:
          type: string
        type:
          type: string
        medium:
          type: string
        difficulty:
          type: string
        marks:
          type: integer
        options:
          type: array
          items:
            type: string
        answer:
          type: string
    ImportResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
    ImportError:
      type: object
      required: [row, message]
      properties:
        row:
          type: integer
        message:
          type: string
...