# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * paperz.pk — school exam paper maker API
 * OpenAPI spec version: 0.1.0
 */
import type { Difficulty } from './difficulty';
import type { Medium } from './medium';
import type { QuestionType } from './questionType';

export interface QuestionUpdate {
  /** @nullable */
  chapterId?: number | null;
  /** @nullable */
  topicId?: number | null;
  type?: QuestionType;
  medium?: Medium;
  difficulty?: Difficulty;
  /** @minimum 0 */
  marks?: number;
  /** @minLength 1 */
  text?: string;
  options?: string[];
  answer?: string;
}