# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * paperz.pk — school exam paper maker API
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  super_admin: 'super_admin',
  school_admin: 'school_admin',
  teacher: 'teacher',
} as const;